+++ src/db/db_user.js
@@ -1,13 +1,13 @@
-/* require('dotenv').config() */
+/* require('dotenv').config({path:'./.env'}) */
 
 const config = {
   db: {
     host: 'localhost',
     user: 'root',
     password: '***REMOVED***',
-    database: 'form',
+    database: 'user',
     connectTimeout: 60000
   },
-  listPerPage: 20,
 };
+
 module.exports = config;