+++ src/server/app.js
@@ -2,13 +2,14 @@ const express = require('express');
 const cors = require('cors');
 const sanitizer = require('perfect-express-sanitizer');
 const formRouter = require('../routes/form');
+const userRouter = require('../routes/user');
 const app = express();
 const port = 3000;
 const log4js = require('log4js');
 log4js.configure('./src/logging/log4js.json');
 const log = log4js.getLogger('startup');
-app.use(cors(), express.json());
 app.use(
+  cors(),
   express.json(),
   express.urlencoded({
     extended: true,
@@ -31,19 +32,19 @@ try {
   }
 }
 
-app.use('/api/send', formRouter);
+app.use('/api/form', formRouter);
+app.use('/api/user', userRouter);
+
 app.use((err, req, res, next) => {
-  log.debug('This is in the app.use function');
+  log.debug('This is in the app.use form function');
   log.error('Something went wrong:', err);
   res.status(err.status || 500);
-  res.render('error', {
-    message: err.message,
-    error: {}
-});
 });
 
 app.listen(port, () => {
   log.debug('This is in the app.listen function');
-  console.log(`Form API listening on http://localhost:${port}`);
+  console.log(`Form API listening on http://localhost:${port}/api/form`);
+  console.log(`User API listening on http://localhost:${port}/api/user`);
   log.info(`Express server listening on port ${port} with pid ${process.pid}`);
 });
+